Fix #80: 3) Limit addOpens to `test` and `run` tasks only
