+++ src/main/java/org/javamodularity/moduleplugin/extensions/RunModuleOptions.java
@@ -2,7 +2,7 @@ package org.javamodularity.moduleplugin.extensions;
 
 import org.gradle.api.Project;
 
-public class RunModuleOptions extends ModuleOptions {
+public class RunModuleOptions extends RuntimeModuleOptions {
 
     public RunModuleOptions(Project project) {
         super(project);
